refactor: removed useless interfaces services
